% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diamond_plot.R
\name{diamond_plot}
\alias{diamond_plot}
\title{Diamond Plot}
\usage{
diamond_plot(df, varx, vary)
}
\arguments{
\item{df}{a data frame}

\item{varx}{a string name for variable x}

\item{vary}{a string name for variable y}
}
\value{
a scatter plot with diamond shaped points
}
\description{
This function produces a ggplot of \code{varx} versus \code{vary}
}
\examples{
df.test <- data.frame(x = rnorm(10), y = rnorm(10))
diamond_plot(df.test, 'x', 'y')
}
